+++ src/main/java/me/basedatos2/pencaucu/dto/team/Teamdto.java
@@ -7,6 +7,6 @@ public class Teamdto {
             String country
     ) {}
     public record DeleteTeamDto(
-            Long id
+            Integer id
     ) {}
 }
